[IMP] point_of_sale: allow product grouping for all uom
In POS, it is hard-codded in such a way that grouping will be done for the
uom category that has database id == 1.

We have solved this issue: we have added a new boolean field "Group Products"
in the uom category form view that will only be visible in developer mode.

If a user chooses this option then products having that uom category will be
grouped in the POS.

Both uom category Unit and Weight have been set to groupable by default.

=========
Technical
=========
 - Removed the hardcoded part of groupable for uom which had database id == 1,
   Now grouping will be done on 'groupable' field of "product.uom", which is
   related to the field of the same name on "product.uom.category".
 - Removed the hardcoded part of is_unit for uom which had database id == 1,
   instead use the name of the uom like it's done elsewhere in the POS...