[IMP] hr_expense: Allow to refuse one or several lines with a reason
Currently, a manager or a accountant can only refuse one expense report with a justification message.

If there are a lot of lines too comment individually or if only one line should be modified, the manager or the accountant can now refused one line and give a reason for it. That will refused the expense report automatically while waiting the employee to make the modifications.

The individually refused lines appear in red in the report form view.