[FIX] stock, purchase: product.packaging_ids in product_template form view
When we activate the setting 'Packaging Methods' in module 'stock'
without having the 'purchase' installed, the product packagings are
not visible in the 'product_template' form view.

This bug is present because the view displaying the product packagings
is defined in the module 'purchase'.

To fix this, we moved the group displaying the packagings from the
product_template inherited view in 'purchase' to the one present in
'stock'.

Bug introduced in rev: https://github.com/odoo/odoo/commit/e945b86d5a61e53a9104ab6cc4ed4a1c060587e4