[IMP] stock: product packagings One2many only relevent fields
When we have the delivery module installed, the tree and form views
representing vendor packagings in the 'product_template' form have
have all the fields added by the delivery module which are not relevent
in this case of vendor packaging.

This is caused by the inheritance of the 'product_packaging_tree_view'
view in the delivery module.

To only show the needed fields in the we have rewritten the the form
and tree views and set them in the context of the 'packaging_ids' field
in the 'product_template' form view.