[refactor] Remove all journals and some sequences
All journals were removed as this is not compliant with
some businesses who actually just manage a few Journals.
The localization will use the default journals and transifex
will be used to do the neccesary changes.

Some ir.sequences were left as this are needed by our
regulator, even though they have no actual impact in Odoo
if not with a third party module.