[IMP] sale_stock: vendor packagin form
When we are creating a new packaging from from a sale order line,
and if the delivery module is installed, the 'product.packaging'
form  shown when we create a new packaging contains all the fields
added by the delivery module.

In this case we want to only show the fields relating a vendor packaging.

So we use the form representing a vendor packaging.