[IMP] product: add volume widget
The volume widget allows you to display a float value stored in the default
uom, in the company's volume unit of measure defined in the general settings.

Both form and kanban widgets have been added, they both go by the name volume.
E.g. (in a form or kanban view):
     <field name="my_volume_field" widget="volume">