TMP WEBSITEPOCALYPSE
[REF] web_editor, website, *: improve module loading + refactoring

The "web_editor.base" module has been split and does not force the
modules which require it to wait for DOM ready anymore. This was indeed
probably slowing loading time, but mainly prevented to use some module
in some contexts (see the LESS editor use in web_studio).
Now the "editor context" can be got thanks to the "web_editor.context"
JS module whose "get" function should be called when the DOM is ready.
The "web_editor.base" still returns its old values, especially the
ready() function which works as before.

If a JS module should wait for the DOM to be ready to be executed, a
new JS module has been created: "web.dom_ready".

Also, the "website.website" module has been split. "website.website"
stills forces the dom to be ready to be executed but does not return
anything anymore, it just initialize some DOM-related stuff. Website
widgets are now grouped in the "website.widgets" module and the topbar
instance can be founded thanks to the "website.menu.instance" module.

This commit also allowed to refactor lots of old code and was the
occasion to lint many JS files. Some other modules have also been
splitted so that widget declaration and widget instanciation occurs
in different module, allowing the declaration code to be executed
before the DOM is ready (which speeds loading time a little and
allows to use the widget in every context).

Note: this refactoring should be improved a lot more than this, but it
is a start towards a better odoo website side.