[FIX] account: avoid crash if no writeoff needed on batch reconcile
Even though the wizard might declare that a writeoff is needed, in
batch mode it might not be the case since the reconciliation is done
in multiple passes for each currency. If no write-off is needed,
no write-off move is created and the code tried to add a boolean
variable to a recordset and raised an apples<>oranges error.

After this fix, if not writeoff move is created, the code simply
launch the full reconcile on the current list of move lines.