[FIX][Accounting] Entry chatter "has been created from : empty"
When a journal entry is created automatically from an invoice
validation, that logs an internal note with the text
"This account entry has been created from:". There should be
a clickable invoice's reference after but there was only an empty space.

The problem come from the method that call the template
"message_origin_link" with the selected invoices as arguments.
The template use the attributes id, _name and name. But for the
invoice object we want to use the number attribute instead of
name. The name of an invoice is empty by default and thus is
rendered as a blank in the template.

This commit modify the template and use the property display_name
instead of name. "display_name" use the number for the invoice and
is still correct for other object.