[ADD] project_timesheet_holidays: when validate leave, generate timesheet
If a vacation project is configured on the current company, when
the HR officer validate the leave, it will generate one analytic line
per day off. The unit_amount will be based on the working calendar of
the employee.
The generated timesheet will be attached to the vacation project of
the company. If this project is not set, no timesheet will be generated.

In the settings, internal project/task of current company can be set. They
will be the default value for project/task of leave type related to this
company during creation. Different project/task can be associate to the
leave type.
If no company set on the leave type (no default value to allow leave type
common to multiple companies), no default value will be set for project/task.
So the HR Officer creating the leave type must be careful to set a project in
the right company to see the generated timesheets.