[FIX] project: pager task count disappeared
Use case when it happens:
* Go to project
* Chose a project
* Go to list view from kanban view
* You can see the pager (item visible/total item)
* Go to another view and come back to list view
* You don't see the pager.

The correct behavior is to never display the pager when the item in
the list view are grouped.

The problem is due to the list view controller. The do search function
is responsible to hide the pager if the item are grouped. But the
objects are asynchronous and the first time we call the list view,
the function do search is called before that the pager rendering.
That implies the pager to be undefined when trying to hide it.

This problem will be deleted with the new list view implementation.

This commit will call the do hide in the pager render function.