[FIX] sales_team : Big target number cannot be used
Use case when it happens:
* Go to crm page
* Set a target value greater than 2147483647(integer max value)
It raises an error integer out of range.
This is problematic for country with a highly inflated currencies.

This happens because target value are stored as Integer in the database.

This commit change field from Integer to Float. The lost in accuracy
for big number does not really matters because it's only used for a
progress bar. The maximum value for a float is 1.7976931348623157e+308
which should be good for all currencies.