[FIX] web_editor, website: restore content menu feature
Commit 3c9e6c89e7207636c1bdab4ed8118c0c1089d43e made big changes to the
link editor dialog... but also broke many feature with it.

Commit 9a1ea6dc08af8b6dc5c5a47277f03f662094d4b2 already fixed some of
them but this commit will fix a bigger one. Indeed, the menu edition
feature was totally broken since the changes. Also, the DOM was messy
which broke form navigation, etc.
This commit also simplify the LESS which was uselessly big.