[FIX] web: reload view with correct params
When reloading a view by clicking on its icon in the view switcher,
some params of the view (e.g. limit) were reset to their original
value. So if the user previously updated that param (e.g. by
editing the pager), these changes were lost.

This rev. ensures that those original params are only given to the
view at its initialization, but not on reload.