[IMP] web_kanban: allow column creation on custom model
In the kanban view, if the default_group_by is a many2one, it's possible
to quick create a new column (which will create a new record in the
relation model).

However, the field `name` was hardcoded during the creation, which means
that a custom model with `x_name` couldn't be correctly created.

This commit allows it by using `name_create` instead.