[IMP] mrp_production: Product in MO filter on possible BOM
Currently, if you create a manufacturing order, you can select
all the product that have a BOM. But some product have BOM that
are archived or that can't be manufactured.

This commit add a constraint in the domain for product field.
The constraint ensure that BOM are not archived nor phantom BOM.