[FIX] pad: no alert at pad creation
When creating a pad before this change these were the steps:

a) beginning of record loading
b) initializing fields
c) pad creation and modification of pad url for record
d) end of record loading
e) url hash is updated to current record
f) alert if the current record is modified

This is an issue because at step "f" some code ascertain there is no
modification and if step "c" happened before (which happen most of the
time) the following alert would always be displayed:

 "The record has been modified, your changes will be discarded. Are you
  sure you want to leave this page ?"

Which is odd to receive just after clicking "Create".

Thus, this fix force step "c" to happen after "f" by waiting that the
record is loaded a the current call stack finish before modifying it
with the new pad url.

opw-702797