[FIX] web: Kanban: record quickcreate: fail handler
When the user quick creates a record in a grouped kanban view, a
name_create RPC is done. This RPC may fail, and there were
several issues with the fail handler.

First, those fail handlers take 2 params: an object describing
the error, and a JQuery event. The event was the first argument
in the handler, so it crashes on event.preventDefault().

Second, when the name_create fails, a FormViewDialog in edit
should be open. Former code opened a SelectCreateDialog.

This commit also ensures that the default values are correctly
given to the form in the dialog, through the context.