[FIX] web: parsing m2o values in domain selector
With the new `field_utils` parsing functions, a m2o value is parsed to
an object with an `id` key and a `display_name` key. This object cannot
be placed in the `<input/>` holding a `DomainLeaf` value in the domain
selector widget (right now, it is showing [Object object] because of
this parsing).
If the value is an object, display its `id` key or nothing if this key
does not exist.