[IMP] calendar: delete recurring events directly from the view
The purpose of these changes is to help the user to delete the recurring events
more easily. Currently if the user want to delete a recurring event he has to:
* Go to the form view
* Uncheck recurring event
* Save
* Delete the event

Now he can directly delete it from the form view. A popup will suggest 3 choices:
* Delete this event
* Delete future events
* Delete all events

This commit creates a new template with 3 buttons and link them to three new events in the js.
The event functions use the server methods to select the needed events and delete them.

Once the events are deleted, the javascript also check for the virtual_id to remove in
the calendar view in order to update it.