[FIX] maintenance: Use a calendar event instead of the cron
Previously the preventive maintenances for equipments use a cron
to generate request. Now it creates a request with a recurring event

When you move a request with a recurring event to a final stage,
it will copy the request and duplicate the linked calendar:
* One calendar with the end date as the scheduled stop date
* Another calendar with a start date equal to the next scheduled date

It will assign the first calendar to the closed request and the second
to the new request.

Modifications on the preventive maintenances for an equipment will
only update the actual and future requests.