[POC] website_portal: stop using session for navigation pager
This is a Proof Of Concept to change thing.

Using document list of the previous 'list' page to compute
the prev/next record to show doesn't work well :
- opening 2 different list pages will make strange
navigation page behavior, as the session content is erased.
- seeing the last record of a list does not allow you to go back to
the first of the list (or the first of next page) when clicking
on 'next'.

prev/next should be computed before rendering. To do so, and minimized
SQL query, I try to use LAG and LEAD sql function. Using direct SQL query
requires to check ir.rule domain, and current domain (filter).
Trying to avoid SQL injection by using ORM method to compute SQL clause.

The limitation of this system will be for ('id', 'in', list_ids)
domain : this will requieres to split domain and execute many sql
queries. That may influence performance for a such small feature.
Since it seems this use case is not manage by the `_search` orm
method, I think we can take the risk.

Future work:
- make the url managed by the template
- the '_get_navigation_urls' should return browse record
(do more stuff in template then)
- extract filter and sortings criteria in method in all
portal modules.